Creates noise module, moves sample to it
