+++ src/lib.rs
@@ -5,6 +5,7 @@ extern crate custom_derive;
 #[macro_use]
 extern crate newtype_derive;
 
+pub mod noise;
 pub mod standard_signals;
 
 custom_derive! {
@@ -106,10 +107,3 @@ impl FrequencyHz {
     }
 }
 
-pub fn sample(rate: FrequencyHz, s: Signal) -> impl Iterator<Item = f64> {
-    (0..).map(move |n: u32| {
-        let sample_period = rate.period();
-        let t = n as f64 * sample_period.0;
-        s.at(TimeSecs(t))
-    })
-}
